递归修改文夹名，递归删除文件

+++ 12.fs/2.deep.js
@@ -1,5 +1,6 @@
 let fs = require('fs');
 let path = require('path');
+// 删除文件为
 function wide(dir,callback){
     // 目录是不是文件
     let arr = [dir];
@@ -9,7 +10,7 @@ function wide(dir,callback){
             let current = arr[--index];
             if(!current) return callback();
             fs.stat(current,function(err,stat){
-                if (stat.isDirectory()) {// 对象描述文件系统目录，则返回 true。
+                if (stat&&stat.isDirectory()) {// 对象描述文件系统目录，则返回 true。
                     fs.rmdir(current,next)
                 }else{
                     fs.unlink(current,next)
@@ -22,7 +23,7 @@ function wide(dir,callback){
         if(index === arr.length) return rmdir()
         let current = arr[index++];
         fs.stat(current,function(err,stat){
-            if(stat.isDirectory()){
+            if (stat&&stat.isDirectory()){
                 fs.readdir(current,function(err,files){ // [b,c]
                     arr = [...arr,...files.map(file=>{
                         return path.join(current,file);
@@ -36,6 +37,29 @@ function wide(dir,callback){
     }
     next();
 }
-wide('./a.txt',function(){
-    console.log('删除成功')
-})
+// wide('./a.txt', function () {
+//     console.log('删除成功')
+// })
+var Dir = "a.txt";
+var Path = "./a";
+deleteFile(Path, Dir);
+
+function deleteFile(paths,dir) {
+	var files = fs.readdirSync(paths);
+	files.forEach(function(file) {
+		if (fs.statSync(paths + "/" + file).isFile()) {
+            var dir = paths + "/" + file;
+            wide(dir, function () {
+                console.log('删除成功')
+                console.log(file, dir);
+            })
+            console.log(file, dir);
+		} else {
+            console.log(paths + "/" + file, "pathFile");
+            var filez = paths + "/" + file;
+            var dirz = filez + dir;
+            deleteFile(filez, dirz);
+		}
+    });
+    
+}
